solana: extract AccountTrie to a new crate
To help reuse the AccountTrie type between different programs (most
notably the IBC module and relayer), introduce solana_trie crate and
move it there.
