+++ Cargo.toml
@@ -7,6 +7,7 @@ rust-version = "1.71.0"
 [workspace]
 members = [
     "common/*",
+    "solana/trie",
     "solana/trie-example",
     "solana/solana-ibc/programs/*",
 ]
@@ -42,6 +43,7 @@ strum = { version = "0.25.0", default-features = false, features = ["derive"] }
 lib = { path = "common/lib" }
 memory = { path = "common/memory" }
 sealable-trie = { path = "common/sealable-trie" }
+solana-trie = { path = "solana/trie" }
 stdx = { path = "common/stdx" }
 
 # dev-dependencies
