+++ solana/trie/Cargo.toml
@@ -1,19 +1,14 @@
 [package]
-name = "solana-trie-example"
+name = "solana-trie"
 authors = ["Michal Nazarewicz <mina86@mina86.com>"]
 version = "0.0.0"
 edition = "2021"
 
-[lib]
-name = "trie"
-crate-type = ["cdylib", "lib"]
-
 [dependencies]
-derive_more.workspace = true
-solana-program.workspace = true
-
 lib.workspace = true
 memory.workspace = true
 sealable-trie.workspace = true
-solana-trie.workspace = true
 stdx.workspace = true
+
+[dev-dependencies]
+solana-program.workspace = true
