+++ solana/trie/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod trie;
+
+pub use sealable_trie::Trie;
+pub use trie::{AccountTrie, DataRef};
